backend/requirements.txt
beautifulsoup4==4.12.3
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-Limiter==3.6.0
google-auth==2.29.0
google-generativeai==0.7.1
gunicorn==22.0.0
langchain==0.2.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==2.0.30
pgvector==0.3.3
Werkzeug==3.0.3